Revert "Revert "handle file upload on upload route""
This reverts commit 2e5cdb09b6263ad3789063cc1679002a4085aec4.

+++ covidml/routes.py
@@ -20,102 +20,27 @@ def home_page():
 @app.route('/upload', methods = ['GET', 'POST'])
 def upload_page():
     form = UploadForm()
+    file_form = UploadFile()
     results = {}
-    
-    try: 
-        if form.validate_on_submit():
-            molecule_id = form.molecule_id.data
-            smiles = [str((form.smiles.data))]
-            model_type = str((form.model_type.data))
-            
-            mask, classifier = load_model(model_type)    
-        
-            activity, confidence, df_original = predict_activity(smiles, mask, classifier)
-            
-            molecular_weight = calculate_molecular_weight(smiles[0])
-                
-            structure_img = smiles_to_image(smiles[0])
-                
-            ad_img, ad_analysis = applicability_domain(molecule_id, df_original)
-                
-            results['molecule_id'] = molecule_id
-            if len(smiles[0]) > 80:
-                results['smiles'] = smiles[0][:80] + '...'
-            else:
-                results['smiles'] = smiles[0]
-            results['mw'] = molecular_weight
-            results['activity'] = activity
-            results['confidence'] = confidence
-            results['ad'] = ad_analysis
-            
-            if not os.path.exists('covidml/static/temp/'):
-                os.mkdir('covidml/static/temp/')
-            
-            if os.path.exists('covidml/static/temp/results.csv'):
-                os.remove('covidml/static/temp/results.csv')
-                
-            df = pd.DataFrame(results, index=[0])
-            df.to_csv('covidml/static/temp/results.csv', index=False)
-                
-                
-            results['model_type'] = model_type
-            results['image'] = structure_img
-            results['adImage'] = ad_img
-            
-    except Exception as e:
-        print(e)
-        flash("Invalid SMILES, enter a valid SMILES to make a prediction", category="danger")
-                
-    return render_template('upload.html', title='Upload', form=form, file_form=UploadFile(), results=results)
+    is_file_tab = False
 
-
-@app.route('/upload_file', methods = ['GET', 'POST'])
-def upload_file_page():
-    file_form= UploadFile()
-    if request.method == 'POST':
-        f = request.files['file']
-        f.save(secure_filename(f.filename))
-        
-        # check if file is a text file
-        if f.filename.split('.')[-1] != 'txt':
-            flash("Invalid file type, please upload a text file", "danger")
-            return render_template('upload_file.html', title='Error', file_form = file_form)
-        
-        # check if file is empty
-        if os.stat(f.filename).st_size == 0:
-            flash("File is empty, please upload a text file with SMILES", "danger")
-            return render_template('upload_file.html', title='Error', file_form = file_form)
-        
-        # check how many non empty lines are in the txt file
-        
-        lines = 0
-        with open(f.filename, 'r') as file:
-            for l in file:
-                if l.strip():
-                    lines += 1
-        
-        for line in range(lines):
-            with open(f.filename, 'r') as file:
-                content = file.readlines()
-            
-            content = content[line].strip().replace('\n', '').split(' ')
-            molecule_id = content[0]
-            smiles = [str(content[1])]
-            model_type = str((file_form.model_type.data))
-        
-            mask, classifier = load_model(model_type)
-        
-            try:
+    if request.args.get('tab') == 'smiles':
+        try: 
+            if form.validate_on_submit():
+                molecule_id = form.molecule_id.data
+                smiles = [str((form.smiles.data))]
+                model_type = str((form.model_type.data))
+                
+                mask, classifier = load_model(model_type)    
             
                 activity, confidence, df_original = predict_activity(smiles, mask, classifier)
-            
+                
                 molecular_weight = calculate_molecular_weight(smiles[0])
-            
+                    
                 structure_img = smiles_to_image(smiles[0])
-            
+                    
                 ad_img, ad_analysis = applicability_domain(molecule_id, df_original)
-            
-                results = {}
+                    
                 results['molecule_id'] = molecule_id
                 if len(smiles[0]) > 80:
                     results['smiles'] = smiles[0][:80] + '...'
@@ -125,29 +50,102 @@ def upload_file_page():
                 results['activity'] = activity
                 results['confidence'] = confidence
                 results['ad'] = ad_analysis
-
+                
                 if not os.path.exists('covidml/static/temp/'):
                     os.mkdir('covidml/static/temp/')
-            
+                
                 if os.path.exists('covidml/static/temp/results.csv'):
                     os.remove('covidml/static/temp/results.csv')
-                
+                    
                 df = pd.DataFrame(results, index=[0])
-                df.to_csv('covidml/static/temp/results.csv', index=False)              
-                                
+                df.to_csv('covidml/static/temp/results.csv', index=False)
+                    
+                    
                 results['model_type'] = model_type
                 results['image'] = structure_img
                 results['adImage'] = ad_img
-            
-            except Exception as e:
-                flash("Invalid SMILES provided in text file", "danger")
-                return render_template('upload_file.html', title='Error', file_form = file_form)
-            
-        # remove uploaded file
-        os.remove(f.filename)
+                
+        except Exception as e:
+            print(e)
+            flash("Invalid SMILES, enter a valid SMILES to make a prediction", category="danger")
 
-        return render_template('upload.html', title='Results', form = UploadForm(), file_form=file_form, results=results)
-    return render_template('upload_file.html', title='Upload File', file_form=file_form)
+    elif request.args.get('tab') == 'file':
+        is_file_tab = True
+        if request.method == 'POST':
+            f = request.files['file']
+            f.save(secure_filename(f.filename))
+            
+            # check if file is a text file
+            if f.filename.split('.')[-1] != 'txt':
+                flash("Invalid file type, please upload a text file", "danger")
+            
+            # check if file is empty
+            if os.stat(f.filename).st_size == 0:
+                flash("File is empty, please upload a text file with SMILES", "danger")
+            
+            # check how many non empty lines are in the txt file
+            
+            lines = 0
+            with open(f.filename, 'r') as file:
+                for l in file:
+                    if l.strip():
+                        lines += 1
+            
+            for line in range(lines):
+                with open(f.filename, 'r') as file:
+                    content = file.readlines()
+                
+                content = content[line].strip().replace('\n', '').split(' ')
+                molecule_id = content[0]
+                smiles = [str(content[1])]
+                model_type = str((file_form.model_type.data))
+            
+                mask, classifier = load_model(model_type)
+            
+                try:
+                
+                    activity, confidence, df_original = predict_activity(smiles, mask, classifier)
+                
+                    molecular_weight = calculate_molecular_weight(smiles[0])
+                
+                    structure_img = smiles_to_image(smiles[0])
+                
+                    ad_img, ad_analysis = applicability_domain(molecule_id, df_original)
+                
+                    results = {}
+                    results['molecule_id'] = molecule_id
+                    if len(smiles[0]) > 80:
+                        results['smiles'] = smiles[0][:80] + '...'
+                    else:
+                        results['smiles'] = smiles[0]
+                    results['mw'] = molecular_weight
+                    results['activity'] = activity
+                    results['confidence'] = confidence
+                    results['ad'] = ad_analysis
+
+                    if not os.path.exists('covidml/static/temp/'):
+                        os.mkdir('covidml/static/temp/')
+                
+                    if os.path.exists('covidml/static/temp/results.csv'):
+                        os.remove('covidml/static/temp/results.csv')
+                    
+                    df = pd.DataFrame(results, index=[0])
+                    df.to_csv('covidml/static/temp/results.csv', index=False)              
+                                    
+                    results['model_type'] = model_type
+                    results['image'] = structure_img
+                    results['adImage'] = ad_img
+                
+                except Exception as e:
+                    flash("Invalid SMILES provided in text file", "danger")
+                
+            # remove uploaded file
+            os.remove(f.filename)
+
+    else:
+        return redirect(url_for('upload_page', tab='smiles'))
+
+    return render_template('upload.html', title='Upload', form=form, file_form=file_form, results=results, is_file_tab=is_file_tab)
 
 
 @app.route('/tutorials')
